Subiendo cambios

+++ src/controller/Session.java
@@ -7,7 +7,7 @@ public class Session {
     public static String USER_NAME;
     public static String USER_ID;
     public static String PAIS_ID;
-    public static int WIDTH_WINDOW;
-    public static int HEIGHT_WINDOW;
+    public static double WIDTH_WINDOW;
+    public static double HEIGHT_WINDOW;
     
 }
